+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## [Unreleased]
 
-- No changes yet.
+- Fix archive and git inputs so that `--path` and `--exclude-path` paths are relative to
+  the `#subdir` rather than the root of the input. This fixes an unintended behavior change
+  that was introduced in `v1.32.0`.
 
 ## [v1.32.0] - 2024-05-16